pyproject.toml
[tool.poetry]
name = "dismake"
version = "1.0.0"
description = "Dismake is a powerful Discord HTTP interactions API wrapper designed for Python developers. Whether you're using Flask, Django, or any other Python web framework, Dismake has you covered."
authors = ["Pranoy Majumdar <officialpranoy@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
        'Development Status :: 1 - Beta',
        'License :: OSI Approved :: MIT License',
        'Intended Audience :: Developers',
        'Natural Language :: English',
        'Operating System :: OS Independent',
        'Programming Language :: Python :: 3.8',
        'Programming Language :: Python :: 3.9',
        'Programming Language :: Python :: 3.10',
        'Programming Language :: Python :: 3.11',
        'Programming Language :: Python :: Implementation :: CPython',
        'Framework :: AsyncIO',
        'Framework :: aiohttp',
        'Topic :: Communications :: Chat',
        'Topic :: Internet :: WWW/HTTP',
        'Topic :: Internet',
        'Topic :: Software Development :: Libraries',
        'Topic :: Software Development :: Libraries :: Python Modules',
        'Topic :: Utilities',
]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.8.5"
colorlog = "^6.7.0"
pynacl = "^1.5.0"
typing-extensions = "^4.7.1"
ruff = "^0.0.287"


[tool.poetry.group.dev.dependencies]
mkdocs = "^1.5.2"
mkdocs-material = "^9.2.5"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}
mypy = "^1.5.1"
mkdocs-awesome-pages-plugin = "^2.9.2"
mkdocs-minify-plugin = "^0.7.1"
black = "^23.7.0"






[tool.poetry.group.test.dependencies]
discord-py = "^2.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
homepage = "https://github.com/Apptic-Development/dismake"
